requirements.txt
nsfw==0.3.2
numpy==1.15.1
pillow==5.2.0
pip==18.0
python-telegram-bot==11.1.0
setuptools==40.2.0
